refactor mongodb connection
